requirements.txt
PyPDF2==3.0.1
openai==1.12.0
gspread==6.0.0
numpy==1.26.4
scikit-learn==1.4.0
requests==2.31.0
google.generativeai
openpyxl
pandas